Changes based on meeting with the Collab Lab
1. Remove colons from all captions so that the captions on figures and
tables match those in List of Figures and List of
Tables. Alternatively, one could add colons to all the items in the
List of Figures, but I chose this solution.

2. Remove margin between the Table of Contents, List of Figures, and
List of Tables so that it matches the margins on the Abstract page and
other preceeding pages. I just eyeballed the length that it needs to
be moved up so I'm not totally sure

3. Add pdf bookmarks for the List of Figures and List of Tables.